+++ doc/changelog.md
@@ -1,7 +1,9 @@
 # ChangeLog
 
 * **0.26.1**
+  - Fix NPE in docker:remove-volumes when no volume configuration is given (#1086)
   - Fix NPE when no networks are configured (#1055)
+  - Fix Base64 encoding for X-Registry-Auth used for Docker authentication (#1084)
 
 * **0.26.1** (2018-07-20)
   - Simple Dockerfile triggered also when only a single run section is given