+++ doc/changelog.md
@@ -6,6 +6,7 @@
   - Fix NPE when no networks are configured (#1055)
   - Fix Base64 encoding for X-Registry-Auth used for Docker authentication (#1084)
   - Fix property configuration based's build detection (#1078)
+  - Introduce container name patterns for naming containers (#931)
 
 * **0.26.1** (2018-07-20)
   - Simple Dockerfile triggered also when only a single run section is given